+++ packages/app/src/runner/tap/commands/health.ts
@@ -1,0 +1,9 @@
+import { defineCommand } from './definition'
+
+export type HealthResult = 'ok'
+
+export const healthCommand = defineCommand({
+  description: 'check that a running Cypress instance is reachable and its tap binding responds',
+  params: [],
+  handler: async (): Promise<HealthResult> => 'ok',
+})
