+++ packages/app/src/runner/tap/commands/index.ts
@@ -1,0 +1,12 @@
+import type { TapCommandDefinition } from './definition'
+import { healthCommand } from './health'
+
+/**
+ * The command registry — the single source of truth for the tap binding.
+ * `getSchema()` serializes the metadata and `TapManager.exec` dispatches by
+ * name, so adding a subcommand is one sibling module (definition, result
+ * types, any seam it needs) plus its entry here.
+ */
+export const tapCommands = {
+  health: healthCommand,
+} satisfies Record<string, TapCommandDefinition>
