+++ packages/app/src/runner/tap/tap-manager.ts
@@ -1,5 +1,6 @@
 import { tapCommands } from './commands'
-import type { TapCommandDefinition } from './commands'
+import { TapCommandError } from './commands/definition'
+import type { TapCommandDefinition } from './commands/definition'
 import { coerceCommandArgs, coerceCommandOptions } from './exec-args'
 import { TAP_PROTOCOL_VERSION } from './contract'
 import type { TapBindingContract, TapExecResult, TapSchema } from './contract'
@@ -21,17 +22,24 @@ const asWireRecord = (value: unknown): Record<string, string> | null => {
   return value as Record<string, string>
 }
 
+/**
+ * The callable surface mounted at `window.__CYPRESS_TAP_BINDING__`. External
+ * callers (the Cypress CLI) invoke the two contract methods over CDP via a
+ * constant `Runtime.callFunctionOn` trampoline, so both must be async,
+ * reachable via `this`, and must take/return only JSON-serializable values
+ * per `./contract`.
+ */
 export class TapManager implements TapBindingContract {
   constructor (private cypressVersion: string) {}
 
-  // Describes this Cypress to the CLI: the protocol version, the Cypress
-  // version, and the full command registry (params/options) it can call via
-  // exec. The CLI reads this to build its interface and validate input.
   async getSchema (): Promise<TapSchema> {
     return {
       protocolVersion: TAP_PROTOCOL_VERSION,
       cypressVersion: this.cypressVersion,
       commands: Object.entries(tapCommands).map(([name, definition]) => {
+        // `satisfies` keeps each entry's narrow literal type, so a command that
+        // omits the optional `options` has no such property to destructure;
+        // widen to the definition shape to read it (defaulting to none).
         const { description, params, options } = definition as TapCommandDefinition
 
         // Snapshot the arrays and their elements so a caller mutating the
@@ -46,11 +54,19 @@ export class TapManager implements TapBindingContract {
     }
   }
 
-  // Runs a single CLI command. Looks up the command, coerces the raw string
-  // args/options into typed values per its schema, then invokes the handler.
-  // Always resolves to a TapExecResult: { ok: true, result } on success, or
-  // { ok: false, code, message } for an unknown command or invalid arguments.
+  /**
+   * The single dispatch entry point. `command`, `args`, and `options` arrive
+   * from the wire as raw strings — `args` and `options` are both maps keyed by
+   * schema name; both are validated and coerced against the registry's
+   * param/option schema before the handler runs. The handler receives the
+   * coerced params object followed by the coerced options object. Both dispatch
+   * failures and domain failures (a handler's `TapCommandError`) come back as
+   * `ok: false`; the handler's own return value is the `ok: true` payload. Any
+   * other throw is a binding bug and propagates to the caller.
+   */
   async exec (command: string, args: Record<string, string> = {}, options: Record<string, string> = {}): Promise<TapExecResult> {
+    // Own-property lookup: `command` is wire input, so an inherited name
+    // like "constructor" must not resolve to a prototype member.
     const definition: TapCommandDefinition | undefined = Object.prototype.hasOwnProperty.call(tapCommands, command)
       ? tapCommands[command as keyof typeof tapCommands]
       : undefined
@@ -90,6 +106,17 @@ export class TapManager implements TapBindingContract {
       return { ok: false, code: 'INVALID_ARGUMENTS', message: coercedOptions.message }
     }
 
-    return { ok: true, result: await definition.handler(coercedArgs.args, coercedOptions.options) }
+    try {
+      return { ok: true, result: await definition.handler(coercedArgs.args, coercedOptions.options) }
+    } catch (err) {
+      // A handler signals a domain failure (no run yet, no such test/spec) by
+      // throwing TapCommandError; surface it as ok: false so the CLI renders it
+      // on stderr and exits non-zero. Any other throw is a real binding bug.
+      if (err instanceof TapCommandError) {
+        return { ok: false, code: err.code, message: err.message }
+      }
+
+      throw err
+    }
   }
 }
